Wed, Dec 28, 2022, 11:44 AM +00:00

+++ filter-more/helpers.c
@@ -40,7 +40,7 @@ for (int i = 0; i <= height - 1 ; i++)
 //setting value 0 to pixel out of image
 for (int a = -1; a <= width; a++)
 {
-    image[-1][a].rgbtBlue = malloc(sizeof(BYTE));
+    image[-1][a].rgbtBlue = malloc(sizeof(int));
     image[-1][a].rgbtBlue = 0;
 
     image[-1][a].rgbtRed = 0;
